+++ src/constant.hpp
@@ -30,20 +30,28 @@ namespace nogdb {
     constexpr unsigned long MAX_DB_SIZE = 1073741824; // 1GB
     constexpr unsigned int MAX_VERSION_CONTROL_SIZE = 128;
     const std::string DB_LOCK_FILE = "/.context.lock";
+
+    constexpr uint16_t INIT_NUM_CLASSES = 4;
     const std::string TB_CLASSES = ".classes";
     const std::string TB_PROPERTIES = ".properties";
     const std::string TB_RELATIONS = ".relations";
     const std::string TB_INDEXES = ".indexes";
+
     const std::string TB_INDEXING_PREFIX = ".index_";
-    constexpr uint16_t UINT16_EM_INIT = 0;
-    const std::string STRING_EM_INIT = ".init";
-    constexpr uint32_t EM_MAXRECNUM = 0;
-    constexpr uint16_t INIT_NUM_CLASSES = 3;
-    constexpr uint16_t INIT_NUM_PROPERTIES = 2;
+
+    constexpr uint16_t INIT_NUM_PROPERTIES = 4;
     constexpr uint16_t CLASS_NAME_PROPERTY_ID = 0;
     const std::string CLASS_NAME_PROPERTY = "@className";
     constexpr uint16_t RECORD_ID_PROPERTY_ID = 1;
     const std::string RECORD_ID_PROPERTY = "@recordId";
+    constexpr uint16_t DEPTH_PROPERTY_ID = 2;
+    const std::string DEPTH_PROPERTY = "@depth";
+    constexpr uint16_t VERSION_PROPERTY_ID = 3;
+    const std::string VERSION_PROPERTY = "@version";
+
+    constexpr uint16_t UINT16_EM_INIT = 0;
+    const std::string STRING_EM_INIT = ".init";
+    constexpr uint32_t EM_MAXRECNUM = 0;
 
 }