+++ src/record.cpp
@@ -19,28 +19,62 @@
  *
  */
 
+#include "constant.hpp"
+#include "validate.hpp"
+
 #include "nogdb_errors.h"
 #include "nogdb_types.h"
 
 namespace nogdb {
 
     Record &Record::set(const std::string &propName, const unsigned char *value) {
-        properties[propName] = Bytes{value, strlen((char *) value)};
+        if (!propName.empty() && propName.at(0) != '@') {
+            properties[propName] = Bytes{value, strlen((char *) value)};
+        }
         return *this;
     }
 
     Record &Record::set(const std::string &propName, const char *value) {
-        properties[propName] = Bytes{reinterpret_cast<const unsigned char *>(value), strlen(value)};
+        if (!propName.empty() && propName.at(0) != '@') {
+            properties[propName] = Bytes{reinterpret_cast<const unsigned char *>(value), strlen(value)};
+        }
         return *this;
     }
 
     Record &Record::set(const std::string &propName, const std::string &value) {
-        properties[propName] = Bytes{static_cast<const unsigned char *>((void *) value.c_str()), strlen(value.c_str())};
+        if (!propName.empty() && propName.at(0) != '@') {
+            properties[propName] = Bytes{static_cast<const unsigned char *>((void *) value.c_str()),
+                                         strlen(value.c_str())};
+        }
         return *this;
     }
 
     Record &Record::set(const std::string &propName, const nogdb::Bytes &b) {
-        properties[propName] = b;
+        if (!propName.empty() && propName.at(0) != '@') {
+            properties[propName] = b;
+        }
+        return *this;
+    }
+
+    Record &Record::setBasicInfo(const std::string &propName, const unsigned char *value) {
+        if (!propName.empty() && propName.at(0) == '@') {
+            properties[propName] = Bytes{value, strlen((char *) value)};
+        }
+        return *this;
+    }
+
+    Record &Record::setBasicInfo(const std::string &propName, const char *value) {
+        if (!propName.empty() && propName.at(0) == '@') {
+            properties[propName] = Bytes{reinterpret_cast<const unsigned char *>(value), strlen(value)};
+        }
+        return *this;
+    }
+
+    Record &Record::setBasicInfo(const std::string &propName, const std::string &value) {
+        if (!propName.empty() && propName.at(0) == '@') {
+            properties[propName] = Bytes{static_cast<const unsigned char *>((void *) value.c_str()),
+                                         strlen(value.c_str())};
+        }
         return *this;
     }
 
@@ -149,19 +183,28 @@ namespace nogdb {
     std::string Record::getText(const std::string &propName) const {
         auto bytes = get(propName);
         if (bytes.empty()) {
-//        throw Error(CTX_NOEXST_PROPERTY, Error::Type::CONTEXT);
             return "";
         } else {
             return bytes.toText();
         }
     }
 
-    void Record::unset(const std::string &className) {
-        properties.erase(className);
+    void Record::unset(const std::string &propName) {
+        if (!propName.empty() && propName.at(0) != '@') {
+            properties.erase(propName);
+        }
+    }
+
+    size_t Record::size() const {
+        auto size = size_t{0};
+        for(auto iter = properties.upper_bound(VERSION_PROPERTY); iter != properties.cend(); ++iter) {
+            ++size;
+        }
+        return size;
     }
 
     bool Record::empty() const {
-        return properties.empty();
+        return size() < 1;
     }
 
     void Record::clear() {