+++ test/test_edge.cpp
@@ -175,22 +175,27 @@ void test_get_edge() {
         auto v2_2 = nogdb::Vertex::create(txn, "persons", r2);
 
         r3.set("time_used", 365U);
-        nogdb::Edge::create(txn, "authors", v1_1, v2_1, r3);
+        auto e1 = nogdb::Edge::create(txn, "authors", v1_1, v2_1, r3);
         r3.set("time_used", 180U);
-        nogdb::Edge::create(txn, "authors", v1_2, v2_1, r3);
+        auto e2 = nogdb::Edge::create(txn, "authors", v1_2, v2_1, r3);
         r3.set("time_used", 430U);
-        nogdb::Edge::create(txn, "authors", v1_3, v2_2, r3);
+        auto e3 = nogdb::Edge::create(txn, "authors", v1_3, v2_2, r3);
 
         auto res = nogdb::Edge::get(txn, "authors");
         auto count = 0;
         for (const auto &it: res) {
             auto &record = it.record;
+            assert(record.getText("@className") == "authors");
+            assert(record.getBigIntU("@version") == 1UL);
             if (count == 0) {
                 assert(record.get("time_used").toIntU() == 365U);
+                assert(record.getText("@recordId") == rid2str(e1.rid));
             } else if (count == 1) {
                 assert(record.get("time_used").toIntU() == 180U);
+                assert(record.getText("@recordId") == rid2str(e2.rid));
             } else if (count == 2) {
                 assert(record.get("time_used").toIntU() == 430U);
+                assert(record.getText("@recordId") == rid2str(e3.rid));
             }
             count++;
         }